fix: prevent circular imports changing some funcs to utils
