+++ statsforecast/_modidx.py
@@ -464,7 +464,6 @@ d = { 'settings': { 'branch': 'main',
                                                                                   'statsforecast/models.py'),
                                       'statsforecast.models._imapa': ('models.html#_imapa', 'statsforecast/models.py'),
                                       'statsforecast.models._intervals': ('models.html#_intervals', 'statsforecast/models.py'),
-                                      'statsforecast.models._naive': ('models.html#_naive', 'statsforecast/models.py'),
                                       'statsforecast.models._optimized_ses_forecast': ( 'models.html#_optimized_ses_forecast',
                                                                                         'statsforecast/models.py'),
                                       'statsforecast.models._predict_mstl_seas': ( 'models.html#_predict_mstl_seas',
@@ -473,11 +472,8 @@ d = { 'settings': { 'branch': 'main',
                                       'statsforecast.models._quantiles': ('models.html#_quantiles', 'statsforecast/models.py'),
                                       'statsforecast.models._random_walk_with_drift': ( 'models.html#_random_walk_with_drift',
                                                                                         'statsforecast/models.py'),
-                                      'statsforecast.models._repeat_val': ('models.html#_repeat_val', 'statsforecast/models.py'),
-                                      'statsforecast.models._repeat_val_seas': ('models.html#_repeat_val_seas', 'statsforecast/models.py'),
                                       'statsforecast.models._seasonal_exponential_smoothing': ( 'models.html#_seasonal_exponential_smoothing',
                                                                                                 'statsforecast/models.py'),
-                                      'statsforecast.models._seasonal_naive': ('models.html#_seasonal_naive', 'statsforecast/models.py'),
                                       'statsforecast.models._seasonal_ses_optimized': ( 'models.html#_seasonal_ses_optimized',
                                                                                         'statsforecast/models.py'),
                                       'statsforecast.models._seasonal_window_average': ( 'models.html#_seasonal_window_average',
@@ -505,4 +501,8 @@ d = { 'settings': { 'branch': 'main',
                                      'statsforecast.theta.thetaforecast': ('theta.html#thetaforecast', 'statsforecast/theta.py'),
                                      'statsforecast.theta.thetamodel': ('theta.html#thetamodel', 'statsforecast/theta.py'),
                                      'statsforecast.theta.thetaupdate': ('theta.html#thetaupdate', 'statsforecast/theta.py')},
-            'statsforecast.utils': {'statsforecast.utils.generate_series': ('utils.html#generate_series', 'statsforecast/utils.py')}}}
+            'statsforecast.utils': { 'statsforecast.utils._naive': ('utils.html#_naive', 'statsforecast/utils.py'),
+                                     'statsforecast.utils._repeat_val': ('utils.html#_repeat_val', 'statsforecast/utils.py'),
+                                     'statsforecast.utils._repeat_val_seas': ('utils.html#_repeat_val_seas', 'statsforecast/utils.py'),
+                                     'statsforecast.utils._seasonal_naive': ('utils.html#_seasonal_naive', 'statsforecast/utils.py'),
+                                     'statsforecast.utils.generate_series': ('utils.html#generate_series', 'statsforecast/utils.py')}}}
